+++ game_defence/src/helpers/line.hpp
@@ -18,8 +18,8 @@ struct LineInfo
 
 struct Line
 {
-  glm::ivec2 p0{ 0, 0 };
-  glm::ivec2 p1{ 0, 0 };
+  glm::vec2 p0{ 0, 0 };
+  glm::vec2 p1{ 0, 0 };
 };
 inline bool
 operator==(const Line& a, const Line& b)
@@ -45,7 +45,7 @@ std::optional<glm::ivec2>
 line_intersection(const glm::ivec2& p0, const glm::ivec2& p1, const glm::ivec2& p2, const glm::ivec2& p3);
 
 glm::ivec2
-get_size_from_line(const Line& l, int line_width = 4);
+get_size_from_line(const Line& l, float line_width = 4);
 
 glm::ivec2
 get_center_from_line(const Line& l);
