+++ game_defence/src/modules/system_change_gun_z_index/helpers.hpp
@@ -5,6 +5,6 @@
 namespace game2d {
 
 void
-set_z_index(entt::registry& r, const entt::entity e, const int index);
+set_z_index(entt::registry& r, const entt::entity e, const float index);
 
 } // namespace game2d
